+++ internal/services/bucket.go
@@ -16,12 +16,12 @@ func (svc *Service) BucketSvc() *BucketService {
 	return &BucketService{Service: svc}
 }
 
-func (svc *BucketService) GetBucket(userID int, bucketID int) (model.Bucket, error) {
+func (svc *BucketService) GetBucket(userID int64, bucketID int64) (model.Bucket, error) {
 	// TODO
 	panic("not implement yet")
 }
 
-func (svc *BucketService) GetUserBuckets(userID int) ([]model.Bucket, error) {
+func (svc *BucketService) GetUserBuckets(userID int64) ([]model.Bucket, error) {
 	resp, err := svc.coordinator.GetUserBuckets(coormsg.NewGetUserBuckets(userID))
 	if err != nil {
 		return nil, fmt.Errorf("get user buckets failed, err: %w", err)
@@ -30,7 +30,7 @@ func (svc *BucketService) GetUserBuckets(userID int) ([]model.Bucket, error) {
 	return resp.Buckets, nil
 }
 
-func (svc *BucketService) GetBucketObjects(userID int, bucketID int) ([]model.Object, error) {
+func (svc *BucketService) GetBucketObjects(userID int64, bucketID int64) ([]model.Object, error) {
 	resp, err := svc.coordinator.GetBucketObjects(coormsg.NewGetBucketObjects(userID, bucketID))
 	if err != nil {
 		return nil, fmt.Errorf("get bucket objects failed, err: %w", err)
@@ -39,7 +39,7 @@ func (svc *BucketService) GetBucketObjects(userID int, bucketID int) ([]model.Ob
 	return resp.Objects, nil
 }
 
-func (svc *BucketService) CreateBucket(userID int, bucketName string) (int, error) {
+func (svc *BucketService) CreateBucket(userID int64, bucketName string) (int64, error) {
 	// TODO 只有阅读了系统操作的源码，才能知道要加哪些锁，但用户的命令可能会调用不止一个系统操作。
 	// 因此加锁的操作还是必须在用户命令里完成，但具体加锁的内容，则需要被封装起来与系统操作放到一起，方便管理，避免分散改动。
 
@@ -61,7 +61,7 @@ func (svc *BucketService) CreateBucket(userID int, bucketName string) (int, erro
 	return resp.BucketID, nil
 }
 
-func (svc *BucketService) DeleteBucket(userID int, bucketID int) error {
+func (svc *BucketService) DeleteBucket(userID int64, bucketID int64) error {
 	// TODO 检查用户是否有删除这个Bucket的权限。检查的时候可以只上UserBucket的Read锁
 
 	mutex, err := reqbuilder.NewBuilder().
